Creating updates execution

+++ app/src/main/java/ru/sibur/android/garbagecollector/Upgrade.java
@@ -1,12 +1,13 @@
 package ru.sibur.android.garbagecollector;
 
 
+import android.content.Context;
 import android.graphics.drawable.VectorDrawable;
 import android.widget.ImageView;
+import android.widget.Toast;
 
 public class Upgrade extends ShopItem {
-    Upgrade(String nameIN, float priceIn){
+    Upgrade(String nameIN, int priceIn){
        super(nameIN, priceIn);
     }
-
 }
